+++ src/test/scala/com.gu.contentapi.client/ClientTest.scala
@@ -4,6 +4,8 @@ import scala.io.Source
 
 trait ClientTest {
 
+  val api = new GuardianContentClient("test")
+
   def loadJson(filename: String): String = {
     Source.fromFile("src/test/resources/templates/" + filename).mkString
   }